Add SEGGER_RTT_TerminalOutBuffer()
Allow writing non-null-terminated buffer to a
specified terminal.

This cannot be done external to this file, because
the variable that tracks the "current" terminal is
a static variable, which limits its name resolution
to this single compilation unit.  Without being
able to read / save the current terminal ID, there
is no safe way to write a buffer to a different
terminal ID and then restore the prior terminal ID.